app/requirements.txt
polars
SQLAlchemy
psycopg2-binary